Version 4.0.0 addresses typing and spec problems:
I wasn't happy with relying on `dom`: the global namespace pollution
kept tripping me up when working on my Node code. So, now I've cloned
the Microsoft typings and adapted them for `node-fetch`.

`node-fetch` had a couple of problems, so I'm now wrapping the final
`Response` object in a `Proxy` to ensure that:

* `Response.type` is present, and
* `Response.body` is a readable stream (see bitinn/node-fetch#210)

The `Response.body` check is a little brittle, as it depends on
inheritance from `stream.Readable` rather than checking for plausible
compatibility to the API, but it's better than having `body` be a
`string` or a `Buffer` depending on how it was initialized.